requirements.txt
Flask==3.1.0
gunicorn
line-bot-sdk==3.16.1